+++ FFmpegDSLAndroid/src/main/java/io/github/vnicius/ffmpegdslandroid/ffmpegcommand/initializer/TimeDurationInitializer.kt
@@ -10,23 +10,23 @@ import io.github.vnicius.ffmpegdslandroid.ffmpegcommand.timeduration.TimeDuratio
  * vinicius.matheus252@gmail.com
  */
 
-fun seconds(second: Float): TimeDuration = TimeDurationBuilder().toSecond(second)
+fun seconds(second: Double): TimeDuration = TimeDurationBuilder().toSecond(second)
 
 fun milliseconds(millisecond: Long): TimeDuration = TimeDurationBuilder().toMillisecond(millisecond)
 
 fun nanoseconds(nanosecond: Long): TimeDuration = TimeDurationBuilder().toNanosecond(nanosecond)
 
 fun hours(hour: Int): TimeDuration =
-    TimeDurationBuilder().toTimePosition(hour, 0, 0f)
+    TimeDurationBuilder().toTimePosition(hour, 0, 0.0)
 
 fun minutes(minute: Int): TimeDuration =
-    TimeDurationBuilder().toTimePosition(0, minute, 0f)
+    TimeDurationBuilder().toTimePosition(0, minute, 0.0)
 
-fun timePosition(second: Float): TimeDuration =
+fun timePosition(second: Double): TimeDuration =
     TimeDurationBuilder().toTimePosition(0, 0, second)
 
-fun timePosition(minute: Int, second: Float): TimeDuration =
+fun timePosition(minute: Int, second: Double): TimeDuration =
     TimeDurationBuilder().toTimePosition(0, minute, second)
 
-fun timePosition(hour: Int, minute: Int, second: Float): TimeDuration =
+fun timePosition(hour: Int, minute: Int, second: Double): TimeDuration =
     TimeDurationBuilder().toTimePosition(hour, minute, second)
